requirements.txt
anyio==2.0.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
filelock==3.0.12
future==0.18.2
html5lib==1.1
idna==2.10
importlib-metadata==3.4.0
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.0
jupyter-server==1.2.2
jupyterlab==3.0.5
jupyterlab-pygments==0.1.2
jupyterlab-server==2.1.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
nbclassic==0.2.6
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
nltk==3.5
notebook==6.2.0
numpy==1.19.5
opencv-python==4.5.1.48
packaging==20.8
pandas==1.2.1
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.14
protobuf==3.14.0
pycparser==2.20
Pygments==2.7.4
pyparsing==2.4.7
pyrsistent==0.17.3
python-crfsuite==0.9.7
python-dateutil==2.8.1
pytz==2020.5
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==21.0.1
regex==2020.11.13
requests==2.25.1
sacremoses==0.0.43
scikit-learn==0.21.3
scipy==1.6.0
selenium==3.141.0
Send2Trash==1.5.0
sentencepiece==0.1.91
seqeval==1.2.2
six==1.15.0
sniffio==1.2.0
soupsieve==2.1
SQLAlchemy==1.3.22
tabulate==0.8.7
terminado==0.9.2
testpath==0.4.4
tokenizers==0.9.3
torch==1.5.1+cpu
torchvision==0.6.1+cpu
tornado==6.1
tqdm==4.56.0
traitlets==5.0.5
transformers==3.5.1
typing-extensions==3.7.4.3
Unidecode==1.1.2
url-normalize==1.4.3
urllib3==1.26.2
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.4.0
